+++ sky-server/src/main/java/com/sky/service/CategoryService.java
@@ -19,5 +19,5 @@ public interface CategoryService {
 
     void update(CategoryDTO categoryDTO);
 
-    List<Category> getByType(long type);
+    List<Category> getByType(Integer type);
 }
